% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city.R
\name{get_city}
\alias{get_city}
\title{Get City Map Layers}
\usage{
get_city(x, r = 1000, verbose = TRUE)
}
\arguments{
\item{x}{city center coordinates}

\item{r}{radi of the extraction}

\item{verbose}{output messages}
}
\value{
A list of layers is returned
}
\description{
Download vairous osm features to create city map layers.
}
\examples{
\dontrun{
bb1 <- osmdata::getbb("Gare Matabiau, 31000 TOULOUSE, France")
lon <- mean(bb1[1, ])
lat <- mean(bb1[2, ])
res <- get_city(c(lon, lat), dist = 700)
}
}
