% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_get.R
\name{om_get}
\alias{om_get}
\title{Get map layers}
\usage{
om_get(x, r = 1000, quiet = FALSE)
}
\arguments{
\item{x}{an sf or sfc object, or a couple of coordinates}

\item{r}{radi of the extraction if x is an sf POINT or a couple of
coordinates}

\item{quiet}{suppress info}
}
\value{
A list of map layers is returned :
\itemize{
\item \emph{zone}, the extraction zone;
\item \emph{urban}, urban areas;
\item \emph{building}, buildings (if the extraction zone area is below 15 km^2);
\item \emph{green}, green spaces;
\item \emph{road}, main roads;
\item \emph{street}, secondary roads;
\item \emph{railway}, railroads (line);
\item \emph{water}, water bodies
If input (\code{x}) uses an unprojected CRS (lon/lat, EPSG:4326), ouput uses
Web Mercator CRS (EPSG:3857), it uses \code{x} CRS otherwise.
}
}
\description{
Download various OpenStreetMap
features to create map layers.
}
\examples{
\dontrun{
bb1 = osmdata::getbb("Gare Matabiau, 31000 TOULOUSE, France")
lon = mean(bb1[1, ])
lat = mean(bb1[2, ])
res = om_get(c(lon, lat), r = 700)
om_map(res)
}
}
