% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_map.R
\name{om_map}
\alias{om_map}
\title{Map layers}
\usage{
om_map(x, title = "Data from OpenStreetMap", theme = "light")
}
\arguments{
\item{x}{a list of map layers}

\item{title}{map title}

\item{theme}{cartographic theme; "light", "dark" or "grey"}
}
\value{
Nothing is returned. A map is displayed.
}
\description{
Display a map of the downloaded layers.
}
\examples{
\dontrun{
r = om_get(c(1.453587,43.611408), 700)
om_map(r, "Matabiau neighbourhood", "light")
}
}
