% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_read.R
\name{om_read}
\alias{om_read}
\title{Import a list of map layers}
\usage{
om_read(filename)
}
\arguments{
\item{filename}{a geopackage file}
}
\value{
A list of map layers is returned.
}
\description{
Read and import a list of layers.
}
\examples{
\dontrun{
r = om_get(c(2.17, 41.39), 500)
om_write(r, "bcn.gpkg")
r = om_read("bcn.gpkg")
om_map(r)
}
}
