% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_write.R
\name{om_write}
\alias{om_write}
\title{Save map layers to a file}
\usage{
om_write(x, filename)
}
\arguments{
\item{x}{a list of map layers produce with \code{om_get}}

\item{filename}{a geopackage file}
}
\value{
The layers are saved to a geopackage file. Nothing is returned.
}
\description{
Write a list a map of layers to a geopackage file.
}
\examples{
\dontrun{
r = om_get(c(2.17, 41.39), 500)
(om_write(r, "bcn.gpkg"))
}
}
